requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
numpy>=1.24.0
piper-tts>=1.2.0
python-multipart>=0.0.6
httpx>=0.25.0